backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pandas
psycopg2-binary
requests
google-generativeai